Update emberAfPostAttributeChangeCallback in examples/ to use uint16_t for size
sync from 705ed321cf861f6a81e0238d9a3ae97a85c2a816

+++ examples/all-clusters-app/ambd/main/CHIPDeviceManager.cpp
@@ -83,7 +83,7 @@ CHIP_ERROR CHIPDeviceManager::Init(CHIPDeviceManagerCallbacks * cb)
     // }
 
     err = Platform::MemoryInit();
-    
+
     SuccessOrExit(err);
 
     // // Register a function to receive events from the CHIP device layer.  Note that calls to
@@ -110,7 +110,7 @@ CHIP_ERROR CHIPDeviceManager::Init(CHIPDeviceManagerCallbacks * cb)
 } // namespace chip
 
 void emberAfPostAttributeChangeCallback(EndpointId endpointId, ClusterId clusterId, AttributeId attributeId, uint8_t mask,
-                                        uint16_t manufacturerCode, uint8_t type, uint8_t size, uint8_t * value)
+                                        uint16_t manufacturerCode, uint8_t type, uint16_t size, uint8_t * value)
 {
     chip::DeviceManager::CHIPDeviceManagerCallbacks * cb =
         chip::DeviceManager::CHIPDeviceManager::GetInstance().GetCHIPDeviceManagerCallbacks();
